lal.series:  Dim instance's .Scale attribute is a float - do not need type cast Original: 8732ff13906ecca1da7733e295d2caac66685111

+++ lal/python/lal/series.py
@@ -98,7 +98,7 @@ def _parse_series(elem, creatorfunc, delta_target_unit_string):
         str(a.Name),
         epoch,
         float(f0.pcdata) * float(f0_unit / inverse_seconds_unit),
-        float(dims[0].Scale) * float(delta_unit / delta_target_unit),
+        dims[0].Scale * float(delta_unit / delta_target_unit),
         sample_unit,
         len(a.array.T)
     )
